+++ test/index.js
@@ -3,6 +3,10 @@ var tests = {
 		'basic': {
 			message: 'supports basic usage'
 		},
+		'basic:clone': {
+			message: 'supports "clone" method',
+			options: { method: 'clone' }
+		},
 		'basic:replace': {
 			message: 'supports "replace" method',
 			options: { method: 'replace' }
@@ -11,6 +15,13 @@ var tests = {
 			message: 'supports "warn" method',
 			options: { method: 'warn' },
 			warning: 2
+		},
+		'vendor': {
+			message: 'ignores vendor prefixes if strict'
+		},
+		'vendor:loose': {
+			message: 'supports vendor prefixes if not strict',
+			options: { strict: false }
 		}
 	}
 };
@@ -32,6 +43,11 @@ Object.keys(tests).forEach(function (name) {
 		t.plan(fixtures.length * 2);
 
 		fixtures.forEach(function (fixture) {
+			var message    = parts[fixture].message;
+			var options    = parts[fixture].options;
+			var warning    = parts[fixture].warning || 0;
+			var warningMsg = message + ' (# of warnings)';
+
 			var baseName   = fixture.split(':')[0];
 			var testName   = fixture.split(':').join('.');
 
@@ -42,24 +58,14 @@ Object.keys(tests).forEach(function (name) {
 			var inputCSS  = fs.readFileSync(inputPath,  'utf8');
 			var expectCSS = fs.readFileSync(expectPath, 'utf8');
 
-			plugin.process(
-				inputCSS,
-				parts[fixture].options
-			).then(function (result) {
+			plugin.process(inputCSS, options).then(function (result) {
 				var actualCSS = result.css;
 
 				if (debug) fs.writeFileSync(actualPath, actualCSS);
 
-				t.equal(
-					actualCSS, expectCSS,
-					parts[fixture].message
-				);
+				t.equal(actualCSS, expectCSS, message);
 
-				t.equal(
-					result.warnings().length,
-					parts[fixture].warning || 0,
-					parts[fixture].message + ' (# of warnings)'
-				);
+				t.equal(result.warnings().length, warning, warningMsg);
 			});
 		});
 	});