+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (2015-10-09)
+
+Added: `strict` option to process prefixed properties.
+Changed: Default `method` is `replace` as there is no spec.
+Updated: Documentation and tests
+
 ## 1.0.0 (2015-10-09)
 
 Added: Initial version